# server/swagger.yaml
openapi: 3.0.3
info:
  title: AVO with Mie - Swagger Page
  description: |-
    This is a server for AVO with Mie website - Website for listening podcast and reading book. 
    
    Some useful links:
    - [The Pet Store repository](https://github.com/ducdatit2002/avowithmie-server)
    - [The source API definition](https://avowithmie-server.onrender.com/api-docs)

  contact:
    email: ducdatit2002@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: https://avowithmie-server.onrender.com/api/
  - url: http://localhost:4000/api/
...
paths:
...
  /books:
    post:
      tags:
        - Books
      summary: Create book
      description: Create a new book entry. Admin only.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to authenticate the admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Book'
                  message:
                    type: string
                    example: 'Book created successfully'
      security:
        - apiKeyAuth: []

    get:
      tags:
        - Books
      summary: Get all books
      description: Retrieve all books.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to access the list of books.
      responses:
        '200':
          description: Books retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
      security:
        - apiKeyAuth: []
  /books/{id}:
    put:
      tags:
        - Books
      summary: Update book by id
      description: Update details of a specific book. Admin only.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed for admin authentication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Book'
                  message:
                    type: string
                    example: 'Update book successfully'
      security:
        - apiKeyAuth: []

    delete:
      tags:
        - Books
      summary: Delete book by id
      description: Remove a book from the database. Admin only.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed for admin authentication.
      responses:
        '200':
          description: Book deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Book deleted successfully'
      security:
        - apiKeyAuth: []
  /books/like/{id}:
    put:
      tags:
        - Books
      summary: Like or unlike a book
      description: User can like or unlike a book.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key required for user authentication.
      responses:
        '200':
          description: Book like status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Added to your liked books'  # Message can vary based on operation
      security:
        - apiKeyAuth: []
  /books/like:
    get:
      tags:
        - Books
      summary: Get all liked books
      description: Retrieve all books liked by the user.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key required for user authentication.
      responses:
        '200':
          description: Liked books retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
      security:
        - apiKeyAuth: []
  /booklists:
    post:
      tags:
        - Booklist
      summary: Create booklist
      description: Create a new booklist. Requires user authentication.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to authenticate the request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookList'
      responses:
        '201':
          description: Booklist created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BookList'
      security:
        - apiKeyAuth: []

    get:
      tags:
        - Booklist
      summary: Get all booklists
      description: Retrieve all booklists. Requires user authentication.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to access the list of booklists.
      responses:
        '200':
          description: Booklists retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookList'
      security:
        - apiKeyAuth: []
  /booklists/{id}:
    put:
      tags:
        - Booklist
      summary: Edit booklist by id
      description: Update details of a specific booklist. Requires user authentication.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to authenticate the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookList'
      responses:
        '200':
          description: Booklist updated successfully
      security:
        - apiKeyAuth: []

    get:
      tags:
        - Booklist
      summary: Get booklist by id
      description: Retrieve a specific booklist by its ID. Requires user authentication.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to authenticate the user.
      responses:
        '200':
          description: Booklist retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BookList'
      security:
        - apiKeyAuth: []

    delete:
      tags:
        - Booklist
      summary: Delete booklist by id
      description: Remove a booklist from the database. Requires user authentication.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed for user authentication.
      responses:
        '200':
          description: Booklist deleted successfully
      security:
        - apiKeyAuth: []
  /booklists/add-book:
    put:
      tags:
        - Booklist
      summary: Add book to booklist
      description: Add a book to an existing booklist. Requires user authentication.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key required to authenticate the request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                booklistId:
                  type: string
                bookId:
                  type: string
      responses:
        '200':
          description: Book added to booklist successfully
      security:
        - apiKeyAuth: []
  /booklists/remove-book:
    put:
      tags:
        - Booklist
      summary: Remove book from booklist
      description: Remove a book from a booklist. Requires user authentication.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key required to authenticate the request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                booklistId:
                  type: string
                bookId:
                  type: string
      responses:
        '200':
          description: Book removed from booklist successfully
      security:
        - apiKeyAuth: []
  /booklists/favourite:
    get:
      tags:
        - Booklist
      summary: Get user's favourite booklists
      description: Retrieve all booklists marked as favourite by the user. Requires user authentication.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to authenticate the request.
      responses:
        '200':
          description: Favourite booklists retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookList'
      security:
        - apiKeyAuth: []
  /booklists/random:
    get:
      tags:
        - Booklist
      summary: Get random booklists
      description: Retrieve a random set of booklists. Requires user authentication.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to authenticate the request.
      responses:
        '200':
          description: Random booklists retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookList'
      security:
        - apiKeyAuth: []
...
components:
  schemas:
...
    Book:
      type: object
      required:
        - title
        - author
        - bookSchema
        - img
      properties:
        title:
          type: string
          example: "Advanced Programming Concepts"
        author:
          type: string
          example: "John Doe"
        bookSchema:
          type: string
          example: "URL to book content"
        img:
          type: string
          description: "URL to book cover image"
          example: "https://example.com/image.jpg"
    BookList:
      type: object
      required:
        - title
        - user
      properties:
        title:
          type: string
          example: "Historical Fiction Favorites"
        user:
          type: string
          description: "User ID of the booklist owner"
        desc:
          type: string
          example: "A collection of top historical fiction books"
        books:
          type: array
          items:
            type: string
          description: "Array of book IDs included in the booklist"
        img:
          type: string
          description: "URL of the booklist cover image"
  securitySchemes:
    auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-auth-token
      description: API key authentication
    adminApiKey:
      type: apiKey
      in: header
      name: x-auth-token
      description: API key authentication for admin endpoints